--- Manager.php
@@ -1,17 +1,0 @@
-<?php
-namespace Projet\Models\admin;
-class Manager
-{
-    protected function dbConnect()
-    {
-
-        try {
-            $bdd = new \PDO('mysql:host=localhost;dbname=projet_biblio;charset=utf8', 'root', '');
-            //$bdd = new \PDO('mysql:host=mysql-chaigwen.alwaysdata.net;dbname=chaigwen_projet_biblio;charset=utf8', 'chaigwen', 'thibaud2901');
-            return $bdd;
-        } catch (Exception $e) {
-
-            die('Erreur : ' . $e->getMessage());
-        }
-    }
-}
