+++ app/models/admin/objets/Manager.php
@@ -7,6 +7,7 @@ class Manager
 
         try {
             $bdd = new \PDO('mysql:host=localhost;dbname=projet_biblio;charset=utf8', 'root', '');
+            //$bdd = new \PDO('mysql:host=mysql-chaigwen.alwaysdata.net;dbname=chaigwen_projet_biblio;charset=utf8', 'chaigwen', 'thibaud2901');
             return $bdd;
         } catch (Exception $e) {
