--- app/public/js/accueil.js
@@ -1,2 +1,0 @@
-
-
