--- app/public/js/monCompte.js
@@ -1,2 +1,0 @@
-
-
