--- app/public/js/nouveaute.js
@@ -1,7 +1,0 @@
-function myFunction(imgs) {
-    var etendreImg = document.getElementsByClassName("etendreImg");
-    var imgText = document.getElementsByClassName("imgtext");
-    etendreImg[0].src = imgs.src;
-    imgText[0].innerHTML = imgs.alt;
-    etendreImg[0].parentElement.style.display = "block";
-  }
