+++ app/views/front/atelier.php
@@ -134,9 +134,8 @@
 
         <!---------- Appel du javascript  / Call of javascript------------>
         <script type="text/javascript" src="./../app/public/js/allJavaScript.js"></script>
-        <script type="text/javascript" src="./../app/public/js/header.js"></script>
         <script type="text/javascript" src="https://www.youtube.com/iframe_api"></script>
-        <script type="text/javascript" src="./../app/public/js/atelier.js"></script>
+        
 
         <!--class active-->
         <script type="text/javascript">
