+++ app/views/front/home.php
@@ -349,8 +349,6 @@
 
         <!---------- Appel du javascript  / Call of javascript------------>
         <script type="text/javascript" src="./../app/public/js/allJavaScript.js"></script>
-        <script type="text/javascript" src="./../app/public/js/header.js"></script>
-        <script type="text/javascript" src="./../app/public/js/accueil.js"></script>
 
         <!--class active-->
         <script type="text/javascript">
