+++ app/views/front/monCompte.php
@@ -243,8 +243,6 @@
 
         <!---------- Appel du javascript  / Call of javascript------------>
         <script type="text/javascript" src="./../app/public/js/allJavaScript.js"></script>
-        <script type="text/javascript" src="./../app/public/js/header.js"></script>
-        <script type="text/javascript" src="./../app/public/js/monCompte.js"></script>
 
     </body>
