+++ app/views/front/pageErreur.php
@@ -59,8 +59,6 @@
 
         <!---------- Appel du javascript  / Call of javascript------------>
         <script type="text/javascript" src="./../app/public/js/allJavaScript.js"></script>
-        <script type="text/javascript" src="./../app/public/js/header.js"></script>
-        <script type="text/javascript" src="./../app/public/js/sliderCoupDeCoeur.js"></script>
 
         <!--class active-->
         <script type="text/javascript">
